backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
djoser==2.2.0
Pillow==10.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0